Patch so that boolean property getters are found first, before a get("foo") is called.
So for a given property foo the order of priority is to look for

* getFoo()
* isFoo()
* get("foo")

So this change just makes boolean getters (isFoo()) of equal priority as other types of getters, like getFoo().
It also has a nice side effect of making the unit tests work as I'd reused Foo for the isSimple() test which also had a get(String) method.


git-svn-id: https://svn-us.apache.org/repos/asf/jakarta/commons/proper/jexl/trunk@136509 13f79535-47bb-0310-9956-ffa450edef68