jexl seems to enjoy producing java.lang.Long from multiplicative expressions, but integral constants are java.lang.Integer.
jexl would only compare an Integer to an Integer for the GT/LT
operators and would throw an exception if one of the operands
was an Integer and one was a Long.  This seemed to be an uneeded
restriction.  So, if either are Long or Integer, the comparison
is made from the resulting longValue() from either to avoid
trimming and to allow comparision between different bitsized
integral values.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn-us.apache.org/repos/asf/jakarta/commons/proper/jexl/trunk@136535 13f79535-47bb-0310-9956-ffa450edef68